Further Work on Spear Game

+++ BrainlessJam/Assets/Scripts/SpearHunting/SpearHuntingConfig.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "SpearHuntingConfig", menuName = "Spear Hunting/Config")]
 public class SpearHuntingConfig : ScriptableObject
 {
-    public float attempts;
+    public int attempts;
     public float launchForce; 
 }
